+++ paper.bib
@@ -9,15 +9,20 @@
   publisher={Oxford University Press}
 }
 
-@article{augustin:2023,
-  title={Filter inference: A scalable nonlinear mixed effects inference approach for snapshot time series data},
-  author={Augustin, David and Lambert, Ben and Wang, Ken and Walz, Antje-Christine and Robinson, Martin and Gavaghan, David},
-  journal={PLOS Computational Biology},
-  volume={19},
-  number={5},
-  pages={e1011135},
-  year={2023},
-  publisher={Public Library of Science San Francisco, CA USA}
+@article{10.1371/journal.pcbi.1011135,
+    doi = {10.1371/journal.pcbi.1011135},
+    author = {Augustin, David AND Lambert, Ben AND Wang, Ken AND Walz, Antje-Christine AND Robinson, Martin AND Gavaghan, David},
+    journal = {PLOS Computational Biology},
+    publisher = {Public Library of Science},
+    title = {Filter inference: A scalable nonlinear mixed effects inference approach for snapshot time series data},
+    year = {2023},
+    month = {05},
+    volume = {19},
+    url = {https://doi.org/10.1371/journal.pcbi.1011135},
+    pages = {1-29},
+    abstract = {Variability is an intrinsic property of biological systems and is often at the heart of their complex behaviour. Examples range from cell-to-cell variability in cell signalling pathways to variability in the response to treatment across patients. A popular approach to model and understand this variability is nonlinear mixed effects (NLME) modelling. However, estimating the parameters of NLME models from measurements quickly becomes computationally expensive as the number of measured individuals grows, making NLME inference intractable for datasets with thousands of measured individuals. This shortcoming is particularly limiting for snapshot datasets, common e.g. in cell biology, where high-throughput measurement techniques provide large numbers of single cell measurements. We introduce a novel approach for the estimation of NLME model parameters from snapshot measurements, which we call filter inference. Filter inference uses measurements of simulated individuals to define an approximate likelihood for the model parameters, avoiding the computational limitations of traditional NLME inference approaches and making efficient inferences from snapshot measurements possible. Filter inference also scales well with the number of model parameters, using state-of-the-art gradient-based MCMC algorithms such as the No-U-Turn Sampler (NUTS). We demonstrate the properties of filter inference using examples from early cancer growth modelling and from epidermal growth factor signalling pathway modelling.},
+    number = {5},
+
 }
 
 @article{SCHUCK:2015,
@@ -64,11 +69,15 @@ author = {Thierry Lavé and Antonello Caruso and Neil Parrott and Antje Walz},
 abstract = {In this review we present ways in which translational PK/PD modeling can address opportunities to enhance probability of success in drug discovery and early development. This is achieved by impacting efficacy and safety-driven attrition rates, through increased focus on the quantitative understanding and modeling of translational PK/PD. Application of the proposed principles early in the discovery and development phases is anticipated to bolster confidence of successfully evaluating proof of mechanism in humans and ultimately improve Phase II success. The present review is centered on the application of predictive modeling and simulation approaches during drug discovery and early development, and more specifically of mechanism-based PK/PD modeling. Case studies are presented, focused on the relevance of M&S contributions to real-world questions and the impact on decision making.}
 }
 
-@article{augustin:20232,
-  title={Simulating clinical trials for model-informed precision dosing: Using warfarin treatment as a use case},
-  author={Augustin, David and Lambert, Ben and Robinson, Martin and Wang, Ken and Gavaghan, David},
-  journal={bioRxiv},
-  pages={2023--07},
-  year={2023},
-  publisher={Cold Spring Harbor Laboratory}
+@article {Augustin2023.07.31.551404,
[…]
