+++ paper.bib
@@ -9,7 +9,7 @@
   publisher={Oxford University Press}
 }
 
-@article{10.1371/journal.pcbi.1011135,
+@article{Augustin:2023,
     doi = {10.1371/journal.pcbi.1011135},
     author = {Augustin, David AND Lambert, Ben AND Wang, Ken AND Walz, Antje-Christine AND Robinson, Martin AND Gavaghan, David},
     journal = {PLOS Computational Biology},
@@ -69,7 +69,7 @@ author = {Thierry Lavé and Antonello Caruso and Neil Parrott and Antje Walz},
 abstract = {In this review we present ways in which translational PK/PD modeling can address opportunities to enhance probability of success in drug discovery and early development. This is achieved by impacting efficacy and safety-driven attrition rates, through increased focus on the quantitative understanding and modeling of translational PK/PD. Application of the proposed principles early in the discovery and development phases is anticipated to bolster confidence of successfully evaluating proof of mechanism in humans and ultimately improve Phase II success. The present review is centered on the application of predictive modeling and simulation approaches during drug discovery and early development, and more specifically of mechanism-based PK/PD modeling. Case studies are presented, focused on the relevance of M&S contributions to real-world questions and the impact on decision making.}
 }
 
-@article {Augustin2023.07.31.551404,
+@article {Augustin:20232,
 	author = {David Augustin and Ben Lambert and Martin Robinson and Ken Wang and David Gavaghan},
 	title = {Simulating clinical trials for model-informed precision dosing: Using warfarin treatment as a use case},
 	elocation-id = {2023.07.31.551404},
@@ -80,4 +80,15 @@ abstract = {In this review we present ways in which translational PK/PD modeling
 	URL = {https://www.biorxiv.org/content/early/2023/08/02/2023.07.31.551404},
 	eprint = {https://www.biorxiv.org/content/early/2023/08/02/2023.07.31.551404.full.pdf},
 	journal = {bioRxiv}
+}
+
+@article{Clerx:2019,
[…]
