### 2023-08-07 #5740 @DavAug
I added the missing DOIs and checked the 'failed' DOI -  it seems to be working just fine. I tried removing 'https://doi.org/' but then the DOI doesn't work anymore.
