### 2023-08-08 #5740 @Kevin-Mattheus-Moerman
@DavAug thanks for this submission. I am the AEiC on this track and here to help with initial steps. I have had a look at your paper. It is in the right ballpark in terms of length, but it is on the small end. I think your `Statement of Need` section can perhaps be expanded. For instance you mention the background of PKPD, and the importance of PKPD models etc., however perhaps you should also mention existing popular open source (and proprietary where applicable) implementations, and compare/contrast  with your work.

You can call: `@editorialbot generate pdf` to update the paper draft.

I'll now proceed to find a handling editor, but you can start on the above already. Note that most editors are currently already preoccupied with many submissions, and many others are on annual leave. Hence it is likely I will need to put this work on a waiting list until an editor becomes available.
