+++ paper.bib
@@ -92,3 +92,44 @@ abstract = {In this review we present ways in which translational PK/PD modeling
  title = {Probabilistic Inference on Noisy Time Series (PINTS)},
  year = {2019}
 }
+
+@article{keizer2013modeling,
+  title={Modeling and simulation workbench for NONMEM: tutorial on Pirana, PsN, and Xpose},
+  author={Keizer, Ron J and Karlsson, MO and Hooker, A},
+  journal={CPT: pharmacometrics \& systems pharmacology},
+  volume={2},
+  number={6},
+  pages={1--9},
+  year={2013},
+  publisher={Wiley Online Library}
+}
+
+@article{hosseini2018gpkpdsim,
+  title={gPKPDSim: a SimBiology{\textregistered}-based GUI application for PKPD modeling in drug development},
+  author={Hosseini, Iraj and Gajjala, Anita and Bumbaca Yadav, Daniela and Sukumaran, Siddharth and Ramanujan, Saroja and Paxson, Ricardo and Gadkar, Kapil},
+  journal={Journal of pharmacokinetics and pharmacodynamics},
+  volume={45},
+  pages={259--275},
+  year={2018},
+  publisher={Springer}
+}
+
+@article{sorzano2021scipion,
+  title={Scipion PKPD: an Open-Source Platform for Biopharmaceutics, Pharmacokinetics and Pharmacodynamics Data Analysis},
+  author={Sorzano, COS and Fonseca-Reyna, Y and de la Cruz-Moreno, MA P{\'e}rez},
+  journal={Pharmaceutical Research},
+  volume={38},
+  number={7},
+  pages={1169--1178},
+  year={2021},
+  publisher={Springer}
+}
+
+@article{rackauckas2020accelerated,
+  title={Accelerated predictive healthcare analytics with pumas, a high performance pharmaceutical modeling and simulation platform},
+  author={Rackauckas, Chris and Ma, Yingbo and Noack, Andreas and Dixit, Vaibhav and Mogensen, Patrick Kofod and Byrne, Simon and Maddhashiya, Shubham and Santiago Calder{\'o}n, Jos{\'e} Bayo{\'a}n and Nyberg, Joakim and Gobburu, Jogarao VS and others},
+  journal={BioRxiv},
+  pages={2020--11},
+  year={2020},
+  publisher={Cold Spring Harbor Laboratory}
+}
