+++ paper.bib
@@ -133,3 +133,14 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   year={2020},
   publisher={Cold Spring Harbor Laboratory}
 }
+
+@article{cole2014maximum,
+  title={Maximum likelihood, profile likelihood, and penalized likelihood: a primer},
+  author={Cole, Stephen R and Chu, Haitao and Greenland, Sander},
+  journal={American journal of epidemiology},
+  volume={179},
+  number={2},
+  pages={252--260},
+  year={2014},
+  publisher={Oxford University Press}
+}
