+++ paper.bib
@@ -144,3 +144,23 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   year={2014},
   publisher={Oxford University Press}
 }
+
+@article{gelman2014understanding,
+  title={Understanding predictive information criteria for Bayesian models},
+  author={Gelman, Andrew and Hwang, Jessica and Vehtari, Aki},
+  journal={Statistics and computing},
+  volume={24},
+  number={6},
+  pages={997--1016},
+  year={2014},
+  publisher={Springer}
+}
+
+@article{augustin2022treatment,
+  title={Treatment response prediction: Is model selection unreliable?},
+  author={Augustin, David and Wang, Ken and Walz, Antje-Christine and Lambert, Ben and Clerx, Michael and Robinson, Martin and Gavaghan, David},
+  journal={bioRxiv},
+  pages={2022--03},
+  year={2022},
+  publisher={Cold Spring Harbor Laboratory}
+}
