+++ paper.bib
@@ -11,7 +11,7 @@
 
 @article{Augustin:2023,
     doi = {10.1371/journal.pcbi.1011135},
-    author = {Augustin, David AND Lambert, Ben AND Wang, Ken AND Walz, Antje-Christine AND Robinson, Martin AND Gavaghan, David},
+    author = {Augustin, David and Lambert, Ben and Wang, Ken and Walz, Antje-Christine and Robinson, Martin and Gavaghan, David},
     journal = {PLOS Computational Biology},
     publisher = {Public Library of Science},
     title = {Filter inference: A scalable nonlinear mixed effects inference approach for snapshot time series data},
