+++ paper.md
@@ -30,7 +30,7 @@ For the sampling from posterior distributions, [Chi](https://chi.readthedocs.io/
 
 PKPD modelling has become an integral part of pharmaceutical research [@SCHUCK:2015; @MORGAN:2018]. In the early phase of drug development, PKPD models help with target and lead identification, and contribute to a (semi-)mechanistic understanding of the relevant pharmacological processes. The modelling results thereafter provide guidance in the transition to the clinical development phase [@LAVE:2016]. During clinical trials, PKPD models help predict the safety, efficacy and treatment response variability of different dosing strategies. More recently, PKPD models are also considered to help with the individualisation of dosing regimens of otherwise difficult-to-administer drugs [@Augustin:20232].
 
-The most widely used programmes for PKPD modelling include NONMEM [@keizer2013modeling], [Monolix](https://lixoft.com/products/monolix/), and Matlab Simbiology [@hosseini2018gpkpdsim]. Other software packages include Scipion PKPD [@sorzano2021scipion], [PoPy](https://product.popypkpd.com/), Pumas [@rackauckas2020accelerated], and a number of [R libraries](https://cran.r-project.org/web/views/Pharmacokinetics.html). While many of these packages provide an extensive toolkit for PKPD modelling, all but Scipion PKPD and the R libraries are closed source, which hinders the ability to further research and extend the existing PKPD modelling methodology.
+The most widely used programmes for PKPD modelling include NONMEM [@keizer2013modeling], [Monolix](https://lixoft.com/products/monolix/), and Matlab Simbiology [@hosseini2018gpkpdsim]. Other software packages include Scipion PKPD [@sorzano2021scipion], [PoPy](https://product.popypkpd.com/), Pumas [@rackauckas2020accelerated], and a number of [R libraries](https://cran.r-project.org/web/views/Pharmacokinetics.html). While many of these packages provide an extensive toolkit for PKPD modelling, all packages are closed source, with the exception of Scipion PKPD and the R libraries. This makes it more challenging to use those software packages for research into novel methodology for PKPD modelling.
 
 [Chi](https://chi.readthedocs.io/en/latest/index.html) is an easy-to-use, open-source Python package for the modelling of PKPD processes. It is targeted at PKPD modellers on all levels of programming expertise. For modellers with an interest in methodological research, [Chi](https://chi.readthedocs.io/en/latest/index.html)'s modular, open source framework makes it easy to extend or replace individual components of PKPD models, as well as investigate their advantages and limitations. The straightforward evaluation of log-likelihoods and log-posteriors, and their sensitivities also facilitates the estimation of parametric and structural uncertainty [@cole2014maximum; @gelman2014understanding; augustin2022treatment].
 
