### 2023-10-04 #5925 @editorialbot
[REVIEW]: Chi: A Python package for treatment response modelling
**Submitting author:** @DavAug (<a href="http://orcid.org/0000-0002-4885-1088">David Augustin</a>)
**Repository:** https://github.com/DavAug/chi
**Branch with paper.md** (empty if default branch): joss-publication
**Version:** 1.0.0
