### 2023-10-09 #5925 @ppxasjsm
@DavAug could you take a look at the missing DOIs flagged by editorial bot and fix them please?
