### 2023-10-09 DavAug/chi#270 @ns-rse
`product` is deprecated as of NumPy 1.25.0, and will be removed in NumPy 2.0. Please use `prod` instead.
Hi,

I'm going through the [review of chi for JOSS](https://github.com/openjournals/joss-reviews/issues/5925) and trying out the examples in the [QuickStart](https://chi.readthedocs.io/en/latest/getting_started/quick_overview.html).

I encountered deprecation warnings with Numpy v1.26.0...

```python
In [13]: # Define log-posterior
    ...: log_prior = pints.ComposedLogPrior(
    ...:     pints.UniformLogPrior(0, 20),  # Initial drug amount
    ...:     pints.UniformLogPrior(0, 20),  # Compartment volume
    ...:     pints.UniformLogPrior(0, 20),  # Elimination rate
    ...:     pints.UniformLogPrior(0, 20)   # Sigma
    ...: )
    ...: log_posterior = chi.LogPosterior(log_likelihood, log_prior)
    ...:
    ...: # Evaluate log-posterior
    ...: score_1 = log_posterior(madeup_parameters)
    ...: score_2 = log_posterior(true_parameters)
<ipython-input-13-919134d94722>:3: DeprecationWarning:

`product` is deprecated as of NumPy 1.25.0, and will be removed in NumPy 2.0. Please use `prod` instead.

<ipython-input-13-919134d94722>:4: DeprecationWarning:

`product` is deprecated as of NumPy 1.25.0, and will be removed in NumPy 2.0. Please use `prod` instead.

<ipython-input-13-919134d94722>:5: DeprecationWarning:

`product` is deprecated as of NumPy 1.25.0, and will be removed in NumPy 2.0. Please use `prod` instead.

<ipython-input-13-919134d94722>:6: DeprecationWarning:

`product` is deprecated as of NumPy 1.25.0, and will be removed in NumPy 2.0. Please use `prod` instead.

In [14]: np.__version__
Out[14]: '1.26.0'
```
