### 2023-10-09 DavAug/chi#270 @DavAug
Thank you @ns-rse for raising this.

I will open a  ticket in pints for this (this is not actually a chi issue, but one for pints - the open source package that I am using here to define the prior distributions).
