+++ paper.bib
@@ -63,7 +63,7 @@ pages = {27-34},
 year = {2016},
 note = {Technology – Translational Pharmacology},
 issn = {1740-6749},
-doi = {https://doi.org/10.1016/j.ddtec.2016.11.005},
+doi = {10.1016/j.ddtec.2016.11.005},
 url = {https://www.sciencedirect.com/science/article/pii/S1740674916300439},
 author = {Thierry Lavé and Antonello Caruso and Neil Parrott and Antje Walz},
 abstract = {In this review we present ways in which translational PK/PD modeling can address opportunities to enhance probability of success in drug discovery and early development. This is achieved by impacting efficacy and safety-driven attrition rates, through increased focus on the quantitative understanding and modeling of translational PK/PD. Application of the proposed principles early in the discovery and development phases is anticipated to bolster confidence of successfully evaluating proof of mechanism in humans and ultimately improve Phase II success. The present review is centered on the application of predictive modeling and simulation approaches during drug discovery and early development, and more specifically of mechanism-based PK/PD modeling. Case studies are presented, focused on the relevance of M&S contributions to real-world questions and the impact on decision making.}
