### 2023-10-09 #5925 @DavAug
> @DavAug could you take a look at the missing DOIs flagged by editorial bot and fix them please?

I fixed the Invalid DOI and manually checked the "Missing" ones. They all link to the correct publication on https://www.doi.org/
