+++ paper.bib
@@ -70,17 +70,15 @@ abstract = {In this review we present ways in which translational PK/PD modeling
 }
 
 @article {Augustin:20232,
-	author = {David Augustin and Ben Lambert and Martin Robinson and Ken Wang and David Gavaghan},
-	title = {Simulating clinical trials for model-informed precision dosing: Using warfarin treatment as a use case},
-	elocation-id = {2023.07.31.551404},
-	year = {2023},
-	doi = {10.1101/2023.07.31.551404},
-	publisher = {Cold Spring Harbor Laboratory},
-	abstract = {Treatment response variability across patients is a common phenomenon in clinical practice. For many drugs this inter-individual variability does not require much (if any) individualisation of dosing strategies. However, for some drugs, including chemotherapies and some monoclonal antibody treatments, individualisation of dosages are needed to avoid harmful adverse events. Model-informed precision dosing (MIPD) is an emerging approach to guide the individualisation of dosing regimens of otherwise difficult-to-administer drugs. Several MIPD approaches have been suggested to predict dosing strategies, including regression, reinforcement learning (RL) and pharmacokinetic and pharmacodynamic (PKPD) modelling. A unified framework to study the strengths and limitations of these approaches is missing. We develop a framework to simulate clinical MIPD trials, providing a cost and time efficient way to test different MIPD approaches. Central for our framework is a clinical trial model that emulates the complexities in clinical practice that challenge successful treatment individualisation. We demonstrate this framework using warfarin treatment as a use case and investigate three popular MIPD methods: 1. neural network regression; 2. deep RL; and 3. PKPD modelling. We find that the PKPD model individualises warfarin dosing regimens with the highest success rate and the highest efficiency: 75.1\% of the individuals display INRs inside the therapeutic range at the end of the simulated trial; and the median time in the therapeutic range (TTR) is 74 \%. In comparison, the regression model and the deep RL model have success rates of 47.9\% and 65.8 \%, and median TTRs of 45 \% and 68 \%. We also find that the MIPD models can attain different degrees of individualisation: the Regression model individualises dosing regimens up to variability explained by covariates; the Deep RL model and the PKPD model individualise dosing regimens accounting also for additional variation using monitoring data. However, the Deep RL model focusses on control of the treatment response, while the PKPD model uses the data also to further the individualisation of predictions.Competing Interest StatementKW is an employee and shareholder of F. Hoffmann-La Roche Ltd. DA, BL, MR and DG have declared that no competing interests exist.},
-	URL = {https://www.biorxiv.org/content/early/2023/08/02/2023.07.31.551404},
-	eprint = {https://www.biorxiv.org/content/early/2023/08/02/2023.07.31.551404.full.pdf},
-	journal = {bioRxiv}
-}
+	AUTHOR={Augustin, David and Lambert, Ben and Robinson, Martin and Wang, Ken and Gavaghan, David},
+  TITLE={Simulating clinical trials for model-informed precision dosing: using warfarin treatment as a use case},
+  JOURNAL={Frontiers in Pharmacology},
+  VOLUME={14},
+  YEAR={2023},
+  URL={https://www.frontiersin.org/articles/10.3389/fphar.2023.1270443},
+  DOI={10.3389/fphar.2023.1270443},
+  ISSN={1663-9812},
+  }
 
 @article{Clerx:2019,
[…]
