### 2023-11-03 #5925 @shahmoradi
Thank you for your efforts. I have checkmarked most review items. I have found a few issues highlighted below:
1. The documentation should ideally start with installation guidelines or pointers to such guidelines (which seem to exist in the readme file of the repository).
2. The example given in the [documentation](https://chi.readthedocs.io/en/latest/getting_started/quick_overview.html) failed with the following error message:
```
C:\ProgramData\Anaconda3\lib\site-packages\numba\np\ufunc\decorators.py in <module>
      1 import inspect
      2
----> 3 from numba.np.ufunc import _internal
      4 from numba.np.ufunc.parallel import ParallelUFuncBuilder, ParallelGUFuncBuilder
      5

SystemError: initialization of _internal failed without raising an exception
```
Please ensure examples or code snippets in the documentation are independent and self-sufficient.
3. Some of the documentation pages are incomplete and should be either removed or completed. Here is an example [broken page](https://chi.readthedocs.io/en/latest/getting_started/log_likelihood.html). Here is another [broken page](https://chi.readthedocs.io/en/latest/getting_started/log_posterior.html).
4. Ideally, there should also be instructions in the repository to run the tests for everyone, novice or expert. But I do see unit-tests and code-coverage reports which is essential.
5. Could you ensure all the requested items in "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" are in place?
