### 2023-11-05 #5925 @DavAug
> Thank you for your efforts. I have checkmarked most review items. I have found a few issues highlighted below:
>
> 1. The documentation should ideally start with installation guidelines or pointers to such guidelines (which seem to exist in the readme file of the repository).
> 2. The example given in the [documentation](https://chi.readthedocs.io/en/latest/getting_started/quick_overview.html) failed with the following error message:
>
> ```
> C:\ProgramData\Anaconda3\lib\site-packages\numba\np\ufunc\decorators.py in <module>
>       1 import inspect
>       2
> ----> 3 from numba.np.ufunc import _internal
>       4 from numba.np.ufunc.parallel import ParallelUFuncBuilder, ParallelGUFuncBuilder
>       5
>
> SystemError: initialization of _internal failed without raising an exception
> ```
>
> Please ensure examples or code snippets in the documentation are independent and self-sufficient. 3. Some of the documentation pages are incomplete and should be either removed or completed. Here is an example [broken page](https://chi.readthedocs.io/en/latest/getting_started/log_likelihood.html). Here is another [broken page](https://chi.readthedocs.io/en/latest/getting_started/log_posterior.html). 4. Ideally, there should also be instructions in the repository to run the tests for everyone, novice or expert. But I do see unit-tests and code-coverage reports which is essential. 5. Could you ensure all the requested items in "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" are in place?

Thank you @shahmoradi for your helpful review and comments.

- [ ] Re 1.: I absolutely agree, and I will add install instructions in the documentation itself. At the moment I am only pointing to the README on the landing page, but I completely understand that this is not ideal.
- [ ] Re 2.: Very curious, I haven't seen this one before. I am a little bit surprised that issues with `numba` in a `conda` environment arise, because my package does not use numba and is currently only distributed with PiPy, so I am surprised that `conda install chi-drm` works in the first place.

Could you, perhaps, provide some more details what you did that spawned this error message (which environment you are using and how you installed the package)?. I will do my best to fix this issue.

- [ ] Re 3.: I completely get your point. These proclamations of extended tutorial documentation are not great. I have been planning to complete the tutorials for a while now, so thank you for pushing me to finish this now. I will complete this in the next couple days and come back to you. Just in case these broken pages discouraged you from scrolling further, the documentation of the API is complete and can be found here: https://chi.readthedocs.io/en/latest/api/index.html
- [ ] Re 4.: That is a helpful recommendation. I will add instructions to run the tests to the README.
- [ ] Re 5.: Yes, absolutely. I will also add a CONTRIBUTION.md to the repository.
