### 2023-11-06 #5925 @DavAug
> Some additional notes I made when reviewing....
>
>
>
> ## Functionality
>
>
>
> ### Installation
>
>
>
> Personally I don't use Debian and so under Arch had to `pacman -Syu sundials`, although I appreciate it can be challenging to cover all use cases, particularly with Linux distributions but perhaps Windows users might benefit from some instructions on how to install external dependencies.
>
>
>
> Installation would perhaps benefit from advising the use of a virtual environment to install the package under but not essential.
>
>
>
> ### Performance
>
>
>
> Claims the package is easier to use than alternatives but personally I'm not familiar with these and did not have the time to investigate or run comparable analyses with the listed alternatives, so I couldn't judge this aspect.
>
>
>
>
>
> ## Documentation
>
>
>
> ### Functionality Documentation
>
>
>
> When I looked I found as @shahmoradi did some of the pages aren't complete, but I see you've already acknowledged this and will address it.
>
>
>
> ### Community Guidelines
>
>
>
> Whilst there is a Contributing section in the `README.md` it is sparse on guidelines (e.g. no mention of the fact Flake8 formatting appears to be applied) and details of how to report issues or seek report. This could be expanded guiding people to create issues if they find problems (not everyone is familiar with GitHub nor how to report problems).
>
>
>
> In this regard the repository could perhaps benefit from having some Issue Templates for reporting of errors which prompts users to provide certain fields such as versions of the package, its dependencies and Python version, copy and pasting the error message, provide input parameters and so forth.  Similarly a Features template would be beneficial.
>
>
>
>

Thank you @ns-rse , really helpful additions! I am in the process of updating the docs and will address your comments in the next couple of days 😊
