### 2023-11-06 #5925 @ns-rse
@DavAug You're welcome.

Certainly not required but [ruff](https://astral.sh/ruff) is an excellent alternative to Flake8, it covers all the rules Flake8 does, is considerably faster and has a [pre-commit](https://github.com/astral-sh/ruff-pre-commit) hook which can be added and integrated with GitHub Actions via [pre-commit.ci](https://pre-commit.ci/).  For more on this see [Research Software Engineering Sheffield - pre-commit : Protecting your future self](https://rse.shef.ac.uk/blog/pre-commit/).
