+++ docs/source/conf.py
@@ -67,7 +67,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'Chi'
-copyright = u'2022, David Augustin'
+copyright = u'2023, David Augustin'
 author = u'David Augustin'
 
 # The version info for the project you're documenting, acts as replacement for
