+++ docs/source/index.rst
@@ -2,6 +2,7 @@
 .. Root of all chi docs
 
 .. _GitHub: https://github.com/DavAug/chi
+.. _Myokit: https://myokit.com
 
 .. module:: chi
 
@@ -20,30 +21,31 @@ which is designed for pharmacokinetic and pharmacodynamic (PKPD) modelling.
 
 The main features of chi are
 
-- Simulation of dose response models for arbitrary dosing regimens, such as pharmacokinetic & pharmacodynamic (PKPD) models, physiology-based pharmacokinetic (PBPK) models, and quantitative systems pharmacology (QSP) models.
-- Inference of model parameters from data (classical or Bayesian).
-- Simulation of the dose response variability in a population (hierarchical models/non-linear mixed effects models).
-- Inference of population parameters from data (classical or Bayesian).
-- Simulation of structured populations, where inter-individual variability can be partly explained by covariates.
-- Inference of model parameters in a structured population from data (classical or Bayesian).
+- Simulation of treatment responses to custom dosing regimens, using pharmacokinetic & pharmacodynamic (PKPD) models, physiology-based pharmacokinetic (PBPK) models, and/or quantitative systems pharmacology (QSP) models.
+- Inference of model parameters from measurements, clinical factors, and/or genetic factors (classical or Bayesian).
+- Simulation of treatment response variability across individuals (hierarchical models/non-linear mixed effects models).
+- Inference of population parameters from measurements, clinical factors, and/or genetic factors (classical or Bayesian).
+- Simulation of structured populations, where inter-individual variability can be partially explained by covariates, such as clinical or genetic factors.
+- Inference of model parameters in a structured population from measurements, clinical factors, and/or genetic factors (classical or Bayesian).
+- Dosing regimen optimisation and model-informed precision dosing (MIPD).
 
-This page provides tutorials and the API documentation for chi.
+This page provides tutorials to illustrate some of chi's functionality, and a detailed API documentation as a complete reference to all of chi's functions and classes.
 
 .. note::
-    Chi is being continuously further developed and improved.
-    So if you find any bugs or have any suggestions for improvement please
+    Chi is being continuously developed and improved.
+    So if you find any bugs or have any suggestions for improvement, please
     don't hesitate to reach out to us and share your feedback.
 
 Install instructions
 --------------------
 
-Installing chi requires two steps: 1. installation of a c-library called sundials; and 2. installation of the python package.
+Installing chi requires two steps: 1. installation of a c-library called sundials; and 2. installation of chi.
 
 Step 1: Installation of sundials
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Chi uses the open source package Myokit to solve ordinary differential equations
-and compute their sensitivities efficiently. Myokit does this using a c-library called sundials.
+Chi uses the open source Python package Myokit_ to solve ordinary differential equations
+and compute their sensitivities efficiently. Myokit_ does this using a c-library called sundials.
 You can install sundials on your computer by entering the below commands in your terminal:
 
 - On Ubuntu, you can execute the below command to install sundials using ``apt-get``:
@@ -77,7 +79,7 @@ Note that you need to install ``chi-drm``, and not ``chi``, to install this pack
 This has the simple reason that the name ``chi`` was already taken in PiPy when we wanted to
 distribute our package.
 
-Now you should be all done and have access to all of chi's functionalities. You can import chi in your python scripts with
+Now you are all done and have access to all of chi's functionality. You can import chi in your python scripts with
 
 .. code-block:: python
 
@@ -85,7 +87,7 @@ Now you should be all done and have access to all of chi's functionalities. You 
 
[…]

[…]
