### 2023-11-18 a1cf5ed
mechanistic model tutorial
