+++ docs/source/getting_started/mechanistic_model.rst
@@ -1,4 +1,5 @@
 .. _SBML: https://sbml.org/
+.. _Myokit: http://myokit.org/
 
 .. currentmodule:: chi
 
@@ -6,29 +7,42 @@
 The mechanistic model
 *********************
 
-In chi, mechanistic model is an umbrella term for time
+In Chi, mechanistic model is an umbrella term for time
 series models describing the dynamics of treatment responses. In general, these models do
 not have to be of a *mechanistic* nature, but in most cases they will have at least
 a mechanistic element to them.
 Popular mechanistic models include for example PKPD models, PBPK models and
 QSP models which are based on mechanistic descriptions of the treatment response.
 
-Mechanistic models can be implemented in chi in two ways: 1. using SBML_ files
+Mechanistic models can be implemented in Chi in two ways: 1. using SBML_ files
 (recommended); and 2. using the :class:`chi.MechanisticModel` interface.
-The :class:`chi.MechanisticModel` way has, perhaps, the
+The :class:`chi.MechanisticModel` interface has, perhaps, the
 lower entry barrier, as it just involves implementing models
-in any Python code you fancy, while the SBML way requires you to learn a small
-amount of potentially unfamiliar SBML syntax. But in the long run, SBML will
-make life easier for you as, in chi, it will automate differentiation,
-sensitivities and dose administration. Below
-we will show how we can use either of those ways to implement a 1-compartment
-PK model.
+using any preferred Python code. This brings an enourmous amount of flexibility
+and, for simple models, gets your model implemented fast. However, it has the
+drawback that sensitivities of the model parameters and
+mechanisms to administer dosing regimens have to implemented manually. It also
+limits the ability to share models.
+
+In comparison, using SBML files to implement your models requires a small
+amount of potentially unfamiliar SBML syntax, and therefore may initially take
+a little longer to get started. However, for complex models, SBML
+simplifies the model implementation and reduces the risk for implementation
+errors. In Chi, SBML will also automate differentiation, the evaluation of parameter
+sensitivities, and the implementation of dose administrations. Another benefit
+of SBML files is that they are programming language-agnostic, meaning that SBML
+files are supported by many simulation softwares facilitating the sharing and
+reimplementation of models without forcing the continued use of Chi. We therefore
+recommend using SBML files to implement your models.
+
+Below we will show how we can use either of those implementation strategies to
+implement a 1-compartment PK model.
 
 **Use case: 1-compartment PK model:**
 
-A 1-compartment PK model describes the absorption, distribution, metabolism and
-elimination of a drug semi-mechanistically using a simple differential equation
-of the form
+A 1-compartment PK model is a semi-mechanistic description of the absorption,
+distribution, metabolism and elimination of a drug using a simple differential
+equation of the form
 
 .. math::
     \frac{\mathrm{d}a}{\mathrm{d}t} = -k_e\, a,
@@ -39,6 +53,248 @@ where :math:`a` is the drug amount in the compartment, :math:`t` is the time,
 :math:`k_e` is the elimination rate of the drug, :math:`c` is the drug
 concentration, :math:`v` is the volume of the compartment and :math:`a_0` is
 the initial drug amount.
+
+Defining mechanistic models using the MechanisticModel interface
+****************************************************************
+
+The simplest way to implement mechanistic models in Chi is to use the
+:class:`chi.MechanisticModel` interface. The :class:`MechanisticModel` is a
+base class for mechanistic models with a small number of mandatory methods that
+you need to implement in order to be able to use your implementation with all
+of Chi's functionality. The base class takes the following form
+(methods that don't have to be implemented are omitted):
+
+.. code-block:: python
+
+    class MechanisticModel(object):
+        def simulate(self, parameters, times):
+            """
[…]
