+++ CONTRIBUTING.md
@@ -1,0 +1,353 @@
+# Contributing to Chi
+
+If you'd like to contribute to Chi (thanks!), please have a look at the [guidelines below](#workflow).
+
+If you're already familiar with our workflow, maybe have a quick look at the [pre-commit checks](#pre-commit-checks) directly below.
+
+## Pre-commit checks
+
+Before you commit any code, please perform the following checks:
+
+- [No style issues](#coding-style-guidelines): `$ flake8`
+- [All tests pass](#testing): `$ python run-tests.py --unit`
+- [The documentation builds](#building-the-documentation): `$ python run-tests.py --doctest`
+
+## Workflow
+
+We use [GIT](https://en.wikipedia.org/wiki/Git) and [GitHub](https://en.wikipedia.org/wiki/GitHub) to coordinate our work. When making any kind of update, we try to follow the procedure below.
+
+### A. Before you begin
+
+1. Create an [issue](https://guides.github.com/features/issues/) where new proposals can be discusssed before any coding is done.
+2. Create a [branch](https://help.github.com/articles/creating-and-deleting-branches-within-your-repository/) of this repo (ideally on your own [fork](https://help.github.com/articles/fork-a-repo/)), where all changes will be made
+3. Download the source code onto your local system, by [cloning](https://help.github.com/articles/cloning-a-repository/) the repository (or your fork of the repository).
+4. [Install](#installation) Chi with the developer options.
+5. [Test](#testing) if your installation worked, using the test script: `$ python run-tests.py --unit`.
+
+You now have everything you need to start making changes!
+
+### B. Writing your code
+
+5. Chi is developed in [Python](https://en.wikipedia.org/wiki/Python_(programming_language)), and makes heavy use of [NumPy](https://en.wikipedia.org/wiki/NumPy) (see also [NumPy for MatLab users](https://docs.scipy.org/doc/numpy/user/numpy-for-matlab-users.html) and [Python for R users](http://blog.hackerearth.com/how-can-r-users-learn-python-for-data-science)).
+6. Make sure to follow our [coding style guidelines](#coding-style-guidelines).
+7. Commit your changes to your branch with useful, descriptive commit messages: Remember these are publically visible and should still make sense a few months ahead in time. While developing, you can keep using the github issue you're working on as a place for discussion. [Refer to your commits](https://stackoverflow.com/questions/8910271/how-can-i-reference-a-commit-in-an-issue-comment-on-github) when discussing specific lines of code.
+8. If you want to add a dependency on another library, or re-use code you found somewhere else, have a look at [these guidelines](#dependencies-and-reusing-code).
+
+### C. Merging your changes with Chi
+
+9. [Test your code!](#testing)
+10. Chi has online documentation at http://chi.readthedocs.io/. To make sure any new methods or classes you added show up there, please read the [documentation](#documentation) section.
+11. When you feel your code is finished, or at least warrants serious discussion, run the [pre-commit checks](#pre-commit-checks) and then create a [pull request](https://help.github.com/articles/about-pull-requests/) (PR) on [Chi's GitHub page](https://github.com/pints-team/chi).
+12. Once a PR has been created, it will be reviewed by any member of the community. Changes might be suggested which you can make by simply adding new commits to the branch. When everything's finished, someone with the right GitHub permissions will merge your changes into Chi's main repository.
+
+## Installation
+
+To install Chi with all developer options, use:
+
+```
+$ git clone https://github.com/pints-team/pints.git
+$ cd pints
+$ pip install -e .[docs]
+```
+
+This will
+
+1. Install all the dependencies for Chi, including the ones for documentation (docs).
+2. Tell Python to use your local Chi files when you use `import chi` anywhere on your system.
+
[…]
