+++ docs/source/getting_started/quick_overview.rst
@@ -6,11 +6,11 @@
 Quick overview
 **************
 
-The quick overview displays some of chi's main features to help you decide
-whether chi is the right software for you. The running example for this
+The quick overview displays some of Chi's main features to help you decide
+whether Chi is the right software for you. The running example for this
 overview will be a 1-compartment pharmacokinetic model (any other deterministic
 time series model would have been an equally good choice, but the 1-compartment
-pharmacokinetic model happens to be predefined in chi's model library;
+pharmacokinetic model happens to be predefined in Chi's model library;
 :class:`chi.library.ModelLibrary`). The model is defined by an ordinary
 differential equation for the drug amount and an algebraic equation for the
 drug concentration
@@ -29,7 +29,7 @@ units for simplicity).
 Simulating a mechanistic model
 ******************************
 
-Using the 1-compartment pharmacokinetic model from chi's model library,
+Using the 1-compartment pharmacokinetic model from Chi's model library,
 simulation of the model simply involves specifying the set of model parameters
 :math:`\psi = (a_0, v, k_e)` and the time points of interest.
 
@@ -49,7 +49,7 @@ The simulation results are of shape ``(n_outputs, n_times)`` which in this case
 is ``(1, 5)`` because we simuated the drug concentration for 5 different time
 points.
 For details on how to implement your own :class:`chi.MechanisticModel` and
-many other details concerning mechanistic models in chi, we refer to
+many other details concerning mechanistic models in Chi, we refer to
 section :doc:`mechanistic_model`.
 
 Visualisation of the simulation
@@ -72,7 +72,7 @@ compartment in intervals of 0.5
 Simulation of measurements
 **************************
 
-Measurements are in chi modelled using a :class:`chi.ErrorModel` on top of the
+Measurements are in Chi modelled using a :class:`chi.ErrorModel` on top of the
 :class:`chi.MechanisticModel` . Formally this takes the form of
 
 .. math::
@@ -96,7 +96,7 @@ aestetic reasons)
    :file: images/1_simulation_2.html
 
 For details on how to implement a :class:`chi.ErrorModel` and
-many other details concerning error models in chi, we refer to section
+many other details concerning error models in Chi, we refer to section
 :doc:`error_model`.
 
 
@@ -104,13 +104,13 @@ Inference of model parameters
 *****************************
 
 While the simulation of mechanistic model outputs and measurements is an
-interesting feature of chi in its own right, the inference of model parameters
+interesting feature of Chi in its own right, the inference of model parameters
 from real-world measurements is arguably even more interesting. In this
 overview, we will use the simulated measurements from above to infer model
 parameters, but in practice the simulated measurements may be straightforwardly
 replaced by real-world measurements.
 
-Inference in chi leverages the fact that the mechanistic model and the
+Inference in Chi leverages the fact that the mechanistic model and the
 error model define a distribution for the modelled measurements :math:`y`.
 In the case of a Gaussian error model, as in the previous example,
 the distribution of the measurements is (yet again) a Gaussian distribution
@@ -129,7 +129,7 @@ parameter values for the measurements
     \log p(\mathcal{D} | \psi , \sigma) =
         \sum _{j=1}^n \log p(y_j | \psi , \sigma , t_j).
 
-In chi the log-likelihood of model parameters can be defined using
+In Chi the log-likelihood of model parameters can be defined using
 :class:`chi.LogLikelihood`. A :class:`chi.LogLikelihood` defines the
 distribution of the measurements using a :class:`chi.MechanisticModel` and a
 :class:`chi.ErrorModel` , and couples the
[…]
