+++ docs/source/index.rst
@@ -19,7 +19,7 @@ Welcome to Chi's documentation!
 **Chi** is an open source Python package hosted on GitHub_,
 which is designed for pharmacokinetic and pharmacodynamic (PKPD) modelling.
 
-The main features of chi are
+The main features of Chi are
 
 - Simulation of treatment responses to custom dosing regimens, using pharmacokinetic & pharmacodynamic (PKPD) models, physiology-based pharmacokinetic (PBPK) models, and/or quantitative systems pharmacology (QSP) models.
 - Inference of model parameters from measurements, clinical factors, and/or genetic factors (classical or Bayesian).
@@ -29,7 +29,7 @@ The main features of chi are
 - Inference of model parameters in a structured population from measurements, clinical factors, and/or genetic factors (classical or Bayesian).
 - Dosing regimen optimisation and model-informed precision dosing (MIPD).
 
-This page provides tutorials to illustrate some of chi's functionality, and a detailed API documentation as a complete reference to all of chi's functions and classes.
+This page provides tutorials to illustrate some of Chi's functionality, and a detailed API documentation as a complete reference to all of chi's functions and classes.
 
 .. note::
     Chi is being continuously developed and improved.
@@ -39,7 +39,7 @@ This page provides tutorials to illustrate some of chi's functionality, and a de
 Install instructions
 --------------------
 
-Installing chi requires two steps: 1. installation of a c-library called sundials; and 2. installation of chi.
+Installing Chi requires two steps: 1. installation of a c-library called sundials; and 2. installation of chi.
 
 Step 1: Installation of sundials
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -63,10 +63,10 @@ You can install sundials on your computer by entering the below commands in your
 
 - On Windows, sundials does not need to be installed manually. Myokit will install sundials automatically.
 
-Step 2: Installation of chi
+Step 2: Installation of Chi
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Chi is distributed with PiPy which means that you can pip install chi with
+Chi is distributed with PiPy which means that you can pip install Chi with
 
 .. code-block:: bash
 
@@ -79,15 +79,15 @@ Note that you need to install ``chi-drm``, and not ``chi``, to install this pack
 This has the simple reason that the name ``chi`` was already taken in PiPy when we wanted to
 distribute our package.
 
-Now you are all done and have access to all of chi's functionality. You can import chi in your python scripts with
+Now you are all done and have access to all of Chi's functionality. You can import chi in your python scripts with
 
 .. code-block:: python
 
     import chi
 
-We hope you enjoy using chi. We are looking forward to seeing which insights you will generate for the pharmaceutical community.
+We hope you enjoy using Chi. We are looking forward to seeing which insights you will generate for the pharmaceutical community.
 
-To get some idea what you can and what you cannot do with chi, we recommend that you have a look at the tutorials on the following pages.
+To get some idea what you can and what you cannot do with Chi, we recommend that you have a look at the tutorials on the following pages.
 The API documentation can be found `here <https://chi.readthedocs.io/en/latest/api/index.html>`_.
 
 .. note::
