+++ docs/source/getting_started/index.rst
@@ -4,9 +4,9 @@ Getting started
 
 .. currentmodule:: chi
 
-This part of the documentation gets you started using chi. Each section will
+This part of the documentation gets you started using Chi. Each section will
 give a brief introduction into the modelling framework and show examples of
-how to implement models in chi. The covered topics include **simulation** and
+how to implement models in Chi. The covered topics include **simulation** and
 **inference** of e.g. ODE models, PKPD models with drug administration and
 non-linear mixed effects models / hierarchical models.
 
