+++ docs/source/getting_started/mcmc_sampling.rst
@@ -1,12 +1,0 @@
-****************************************************
-Inferring posterior distributions with MCMC sampling
-****************************************************
-
-.. currentmodule:: chi
-
-This site is still under construction. In the mean time, we encourage you to
-have a look at the detailed API documentation
-
-.. autosummary::
-
-    chi.SamplingController
