### 2023-11-23 3b7959b
fitting models draft
