+++ docs/source/getting_started/code/3_fitting_models_2.py
@@ -1,0 +1,377 @@
+import argparse
+
+
+# Set up argument parsing, so plotting and exports can be disabled for
+# testing.
+parser = argparse.ArgumentParser(
+    description='Run example scripts for chi docs.',
+)
+parser.add_argument(
+    '--test',
+    action='store_true',
+    help='Run testing version of script which ignores plotting.',)
+
+# Parse!
+args = parser.parse_args()
+
+
+import os
+
+import chi
+import numpy as np
+import pandas as pd
+import pints
+import plotly.graph_objects as go
+
+# 3
+import pints
+
+
+# Define mechanistic model
+directory = os.path.dirname(__file__)
+filename = os.path.join(directory, 'one_compartment_pk_model.xml')
+model = chi.PKPDModel(sbml_file=filename)
+model.set_administration(compartment='global', direct=False)
+model.set_outputs(['global.drug_concentration'])
+
+# Define error model
+error_model = chi.LogNormalErrorModel()
+
+# Define data
+directory = os.path.dirname(__file__)
+data = pd.read_csv(directory + '/dataset_1.csv')
+
+# Define prior distribution
+log_prior = pints.ComposedLogPrior(
+    pints.GaussianLogPrior(mean=10, sd=2),           # absorption rate
+    pints.GaussianLogPrior(mean=6, sd=2),            # elimination rate
+    pints.LogNormalLogPrior(log_mean=0, scale=1),    # volume of distribution
+    pints.LogNormalLogPrior(log_mean=-2, scale=0.5)  # scale of meas. distrib.
+)
+
+# Define log-posterior using the ProblemModellingController
+problem = chi.ProblemModellingController(
+    mechanistic_model=model, error_models=[error_model])
+problem.set_data(
+    data=data,
+    output_observable_dict={'global.drug_concentration': 'Drug concentration'}
+)
+problem.fix_parameters(name_value_dict={
+    'dose.drug_amount': 0,
+    'global.drug_amount': 0,
+})
+problem.set_log_prior(log_prior=log_prior)
+log_posterior = problem.get_log_posterior()
+
+
+# 4
+# Run MCMC algorithm
+n_iterations = 20000
+controller = chi.SamplingController(log_posterior=log_posterior, seed=1)
+controller.set_n_runs(n_runs=3)
+controller.set_parallel_evaluation(False)
+controller.set_sampler(pints.HaarioBardenetACMC)
+samples = controller.run(n_iterations=n_iterations, log_to_screen=True)
+
+# 5
+from plotly.colors import qualitative
+from plotly.subplots import make_subplots
+
+
+# Plot results
+fig = make_subplots(
+    rows=4, cols=2, vertical_spacing=0.15, horizontal_spacing=0.1)
+
+# Plot traces and histogram of parameter
+fig.add_trace(
+    go.Scatter(
+        name='Run 1',
+        x=np.arange(1, n_iterations+1),
+        y=samples['dose.absorption_rate'].values[0],
+        mode='lines',
+        line_color=qualitative.Plotly[2],
+    ),
+    row=1,
+    col=1
+)
+fig.add_trace(
+    go.Scatter(
+        name='Run 2',
+        x=np.arange(1, n_iterations+1),
+        y=samples['dose.absorption_rate'].values[1],
+        mode='lines',
+        line_color=qualitative.Plotly[1],
+    ),
+    row=1,
+    col=1
+)
+fig.add_trace(
+    go.Scatter(
+        name='Run 3',
+        x=np.arange(1, n_iterations+1),
+        y=samples['dose.absorption_rate'].values[2],
+        mode='lines',
+        line_color=qualitative.Plotly[0],
+    ),
[…]
