+++ docs/source/getting_started/fitting_models_to_data.rst
@@ -1,5 +1,8 @@
 .. currentmodule:: chi
 
+.. _Dataset_1 : https://github.com/DavAug/chi/blob/main/docs/source/getting_started/data/dataset_1.csv
+.. _Pints: https://github.com/pints-team/pints
+
 **********************************
 Fitting mechanistic models to data
 **********************************
@@ -7,27 +10,373 @@ Fitting mechanistic models to data
 In the previous tutorial, :doc:`mechanistic_model`, we have seen how we can
 implement and simulate treatment response models in Chi. For example, using the same
 1-compartment PK model from before, ``one_compartment_pk_model.xml``, we can simulate
-the time course of drug concentration levels following repeated bolus
+the time course of drug concentration levels following repeated dose
 adminstrations
 
 .. literalinclude:: code/3_fitting_models_1.py
-    :lines: 18-51
+    :lines: 18-52
 
 .. raw:: html
    :file: images/3_fitting_models_1.html
 
 This ability to simulate treatment response is pretty cool in its own right,
-but, at this point, our model is completely made up and has nothing to do with
+but, at this point, our model has little to do with
 real treatment responses. If our goal is to describe *real* treatment
 responses, we need to somehow connect our model to reality.
 
 The most common approach to relate models to real treatment responses is to
-compare the model predictions to measurements.
+compare the model predictions to measurements. Below, we have prepared an example
+dataset with drug concentration measurements. These drug concentrations were
+measured after repeatedly adminstering 2 mg of a drug every 24 hours.
+You can download the dataset from the following link, if you would like to
+follow this tutorial: Dataset_1_.
 
 .. csv-table:: Drug concentration measurements
-   :file: data/3_fitting_models_1.csv
+   :file: code/dataset_1.csv
    :widths: 4, 12, 12, 12, 12, 12, 12, 12, 12
    :header-rows: 1
+
+The dataset contains one column that identifies the measured individual (``ID``),
+two columns that specify the time of a measurement or a dose administration
+(``Time`` and ``Time unit``), three columns that specify the measured values
+(``Observable``, ``Value``, ``Observable unit``), and three columns that specify
+the dose administrations (``Dose``, ``Duration``, ``Dose unit``).
+
+If we download the file and save it in the same directory as the Python script,
+we can visualise the measurements by executing the below script
+
+.. literalinclude:: code/3_fitting_models_1.py
+    :lines: 82-101
+
+.. raw:: html
+   :file: images/3_fitting_models_2.html
+
+The figure shows that the treatment response dynamics indicated by the measurements
+is similar to treatment response simulated by the one-compartment PK model above.
+But looking more closely at the magnitude of the values,
+it appears that the measured values are much smaller
+than the simulated ones. We can therefore conclude that, at this point, our
+model does not provide an accurate description of the measured treatment
+response.
+
+To find a better description of the treatment response, we have two options:
+1. we can try to find parameters values that make the model output describe the
+measurements more closely; or 2. we can define a new mechanistic model and see
+whether this new model is able to describe the measurements better. This tutorial
+will be about how we can find better model parameters.
+
+Estimating model parameters from data: Background
+*************************************************
+
+Before we can try to find parameter values that most closely describe the treatment response
+measurements, we first need to agree on what we mean by
+"*most closely*" for the relationship between the mechanistic model output and the measurements.
+The most naive way to define this notion of closeness is to use the distance
+between measurements and model outputs,
+i.e. the difference between the measured value and the
+simulated value. If we used
[…]
