+++ docs/source/getting_started/code/3_fitting_models_2.py
@@ -84,41 +84,9 @@ fig = make_subplots(
 
 # Plot traces and histogram of parameter
 fig.add_trace(
-    go.Scatter(
-        name='Run 1',
-        x=np.arange(1, n_iterations+1),
-        y=samples['dose.absorption_rate'].values[0],
-        mode='lines',
-        line_color=qualitative.Plotly[2],
-    ),
-    row=1,
-    col=1
-)
-fig.add_trace(
-    go.Scatter(
-        name='Run 2',
-        x=np.arange(1, n_iterations+1),
-        y=samples['dose.absorption_rate'].values[1],
-        mode='lines',
-        line_color=qualitative.Plotly[1],
-    ),
-    row=1,
-    col=1
-)
-fig.add_trace(
-    go.Scatter(
-        name='Run 3',
-        x=np.arange(1, n_iterations+1),
-        y=samples['dose.absorption_rate'].values[2],
-        mode='lines',
-        line_color=qualitative.Plotly[0],
-    ),
-    row=1,
-    col=1
-)
-fig.add_trace(
     go.Histogram(
         name='Posterior samples',
+        legendgroup='Group 1',
         x=samples['dose.absorption_rate'].values[:, n_iterations//2::(n_iterations//2)//1000].flatten(),
         histnorm='probability density',
         showlegend=True,
@@ -126,49 +94,50 @@ fig.add_trace(
     marker_color=qualitative.Plotly[4],
     ),
     row=1,
+    col=1
+)
+fig.add_trace(
+    go.Scatter(
+        name='Run 1',
+        legendgroup='Group 2',
+        x=np.arange(1, n_iterations+1),
+        y=samples['dose.absorption_rate'].values[0],
+        mode='lines',
+        line_color=qualitative.Plotly[2],
+    ),
+    row=1,
+    col=2
+)
+fig.add_trace(
+    go.Scatter(
+        name='Run 2',
+        legendgroup='Group 3',
+        x=np.arange(1, n_iterations+1),
+        y=samples['dose.absorption_rate'].values[1],
+        mode='lines',
+        line_color=qualitative.Plotly[1],
+    ),
+    row=1,
+    col=2
+)
+fig.add_trace(
+    go.Scatter(
+        name='Run 3',
+        legendgroup='Group 4',
+        x=np.arange(1, n_iterations+1),
+        y=samples['dose.absorption_rate'].values[2],
+        mode='lines',
+        line_color=qualitative.Plotly[0],
+    ),
+    row=1,
     col=2
 )
 
 # Plot traces and histogram of parameter
 fig.add_trace(
-    go.Scatter(
-        name='Run 1',
-        x=np.arange(1, n_iterations+1),
-        y=samples['global.elimination_rate'].values[0],
-        mode='lines',
-        line_color=qualitative.Plotly[2],
-        showlegend=False
-    ),
-    row=2,
-    col=1
-)
-fig.add_trace(
-    go.Scatter(
-        name='Run 2',
-        x=np.arange(1, n_iterations+1),
-        y=samples['global.elimination_rate'].values[1],
-        mode='lines',
-        line_color=qualitative.Plotly[1],
-        showlegend=False
-    ),
-    row=2,
-    col=1
-)
-fig.add_trace(
-    go.Scatter(
-        name='Run 3',
-        x=np.arange(1, n_iterations+1),
-        y=samples['global.elimination_rate'].values[2],
-        mode='lines',
-        line_color=qualitative.Plotly[0],
-        showlegend=False
-    ),
-    row=2,
-    col=1
-)
-fig.add_trace(
     go.Histogram(
         name='Posterior samples',
+        legendgroup='Group 1',
[…]
