+++ docs/source/getting_started/modelling_IIV.rst
@@ -1,18 +1,0 @@
-.. currentmodule:: chi
-
-**************************************
-Modelling inter-individual variability
-**************************************
-
-This site is still under construction. In the mean time, we encourage you to
-have a look at the detailed API documentation
-
-.. autosummary::
-
-    chi.PopulationModel
-    chi.ComposedPopulationModel
-    chi.GaussianModel
-    chi.LogNormalModel
-    chi.PooledModel
-    chi.TruncatedGaussianModel
-    chi.ReducedPopulationModel
