### 2023-11-26 720e7a7
clean up for doc testing
