+++ docs/source/getting_started/mechanistic_model.rst
@@ -930,8 +930,8 @@ elimination rate is negative, in contrast to the positive sensitivity with
 respect to the initial drug amount. As a result, larger elimination rates
 negatively impact drug conentration levels.
 
-This concludes this tutorial. If you have any feedback to improve or would like
-to report any typos, mistakes or bugs, please do reach out to us, for example
+This concludes this tutorial. If you have any feedback or suggestions for
+improvement, or would like to report any typos, mistakes or bugs, please do reach out to us, for example
 by creating an Issue_. We are looking
 forward to hearing from you!
 

[…]
