### 2023-11-26 #5925 @DavAug
Hi @ns-rse @shahmoradi @ppxasjsm @Kevin-Mattheus-Moerman

I hope you had a great weekend and thank you for your patience :)

I have completed the cleaning up the documentation page and tried to address your comments above. In particular:
1. I have added install instructions to the landing page of the documentation: https://chi.readthedocs.io . These instructions detail how to install the package using pip on Windows, MacOS and Ubuntu. I chose these operating systems, because I can test the installs on these operating systems with GitHub workflows. (This means I have not included the install instructions that you kindly provided @ns-rse . I hope that is ok).
3. I have extended the tutorials quite substantially to demonstrate in more detail how to use Chi and to highlight the benefits of using Chi, see Getting started section in the documentation. I have workflows in place that test these tutorials, so the examples should work if chi is correctly installed, at least on the tested operating systems. (My unit test workflows run the tests on Python version 3.9, 3.10 and 3.11.)
4. I have removed any pointers to tutorials that may be added in the future. So there are no "broken" pages in the documentation left.
5. I have added an extensive CONTRIBUTING.md which hopefully encourages the community to contribute to Chi and simplifies the onboarding.

The one point that remains unaddressed are your challenges with installing the package, @shahmoradi . Would you, perhaps, be able to try installing the package again following the install instructions on https://chi.readthedocs.io . I hope that this will resolve the problem :)

I hope you have a nice rest of the day and best wishes,
David
