### 2023-12-04 794b602
doc tweaks
