+++ docs/source/getting_started/fitting_models_to_data.rst
@@ -1,6 +1,6 @@
 .. currentmodule:: chi
 
-.. _Dataset_1 : https://github.com/DavAug/chi/blob/main/docs/source/getting_started/data/dataset_1.csv
+.. _Dataset_1 : https://github.com/DavAug/chi/blob/main/docs/source/getting_started/code/dataset_1.csv
 .. _Pints: https://github.com/pints-team/pints
 .. _ArViz: https://python.arviz.org/en/stable/
 .. _Issue: https://github.com/DavAug/chi/issues
@@ -10,10 +10,10 @@ Fitting mechanistic models to data
 **********************************
 
 In the previous tutorial, :doc:`mechanistic_model`, we have seen how we can
-implement and simulate treatment response models in Chi. For example, using the same
-1-compartment PK model from before, ``one_compartment_pk_model.xml``, we can simulate
+implement and simulate treatment response models in Chi. For example, we can simulate
 the time course of drug concentration levels following repeated dose
-adminstrations
+adminstrations as follows, using the same
+1-compartment PK model as before
 
 .. literalinclude:: code/3_fitting_models_1.py
     :lines: 112-146
@@ -21,28 +21,27 @@ adminstrations
 .. raw:: html
    :file: images/3_fitting_models_1.html
 
-This ability to simulate treatment response is pretty cool in its own right,
+This ability to simulate treatment responses is pretty cool in its own right,
 but, at this point, our model has little to do with
-real treatment responses. If our goal is to describe *real* treatment
+real treatment responses. Assuming that our goal is to describe *real* treatment
 responses, we need to somehow connect our model to reality.
 
 The most common approach to relate models to real treatment responses is to
 compare the model predictions to measurements. Below, we have prepared an example
 dataset with drug concentration measurements. These drug concentrations were
 measured after repeatedly adminstering 2 mg of a drug every 24 hours.
-You can download the dataset from the following link, if you would like to
-follow this tutorial: Dataset_1_.
+You can download the dataset from the following link: Dataset_1_.
 
 .. csv-table:: Drug concentration measurements
    :file: code/dataset_1.csv
    :widths: 4, 12, 12, 12, 12, 12, 12, 12, 12
    :header-rows: 1
 
-The dataset contains one column that identifies the measured individual (``ID``),
-two columns that specify the time of a measurement or a dose administration
-(``Time`` and ``Time unit``), three columns that specify the measured values
+The dataset contains one column that identifies measured individuals (``ID``),
+two columns that specify measurement times and dose administration times
+(``Time`` and ``Time unit``), three columns that specify measured values
 (``Observable``, ``Value``, ``Observable unit``), and three columns that specify
-the dose administrations (``Dose``, ``Duration``, ``Dose unit``).
+dose administrations (``Dose``, ``Duration``, ``Dose unit``).
 
 If we download the file and save it in the same directory as the Python script,
 we can visualise the measurements by executing the below script
@@ -62,30 +61,30 @@ model does not provide an accurate description of the measured treatment
 response.
 
 To find a better description of the treatment response, we have two options:
-1. we can try to find parameters values that make the model output describe the
-measurements more closely; or 2. we can define a new mechanistic model and see
+1. we can try to find parameter values that improves the proximity of the
+model output to the measurements; or 2. we can define a new mechanistic model and see
 whether this new model is able to describe the measurements better. This tutorial
-will be about how we can find better model parameters.
+will be about the former and detail how we can find better model parameters.
 
 Estimating model parameters from data: Background
 *************************************************
 
-Before we can try to find parameter values that most closely describe the treatment response
-measurements, we first need to agree on what we mean by
+Before we can try to find parameter values that describe the observed
+treatment response most closely, we first need to agree on what we mean by
 "*most closely*" for the relationship between the mechanistic model output and the measurements.
[…]
