### 2023-12-06 #5925 @DavAug
Hi and good morning @shahmoradi ,

I just wanted to kindly follow up on the above and ask whether you were able to successfully install Chi now?

I hope you have a nice day and best wishes,
David
