### 2023-12-13 #5925 @ns-rse
Morning @DavAug ,

Thanks again for the nudge. Had a look through the contributing guidelines are a great addition, thanks for adding those. There is a clear statement of need too and so I've ticked off the last couple of items on my check list.

If you've ever time to work on this I can highly recommend `pre-commit` and `pre-commit.ci` the later of which integrates well with a GitHub action and catches instances where contributors might not have applied Flake8 locally before commits (I added some links to more information in a [previous comment](https://github.com/openjournals/joss-reviews/issues/5925#issuecomment-1795198392) that may be useful).

Have fun and all the best,

@ns-rse
