### 2023-12-15 DavAug/chi#273 @DavAug
Investigate alternative to flask
Below comments from JOSS review:

Certainly not required but [ruff](https://astral.sh/ruff) is an excellent alternative to Flake8, it covers all the rules Flake8 does, is considerably faster and has a [pre-commit](https://github.com/astral-sh/ruff-pre-commit) hook which can be added and integrated with GitHub Actions via [pre-commit.ci](https://pre-commit.ci/). For more on this see [Research Software Engineering Sheffield - pre-commit : Protecting your future self](https://rse.shef.ac.uk/blog/pre-commit/).

If you've ever time to work on this I can highly recommend pre-commit and pre-commit.ci the later of which integrates well with a GitHub action and catches instances where contributors might not have applied Flake8 locally before commits (I added some links to more information in a https://github.com/openjournals/joss-reviews/issues/5925#issuecomment-1795198392 that may be useful).
