### 2023-12-19 #5925 @DavAug
> Thanks @DavAug, for your reminder and response. The error still exists, but given your package does not use `numba`, I take it as an issue for me to resolve. I have checkmarked the rest of the items that you have addressed.

Super, thank you so much!
