### 2024-01-02 #5925 @DavAug
Hi @ppxasjsm @Kevin-Mattheus-Moerman and happy New Year! 😊

I just wanted to kindly follow up on this and ask whether there are any remaining steps on my side to conclude this process, given that both reviewers are now satisfied with the submission?

I hope you have a great day and best wishes,
David
