### 2024-01-14 #5925 @DavAug
Super thank and no problem at all. Let me copy-paste the list, so I can tick it off while working on it :)

 - [ ] add missing references
 - [ ] Double check your author affiliation including ORCID
