### 2024-01-14 0a21834
Add references and tweaks
