+++ paper.bib
@@ -6,7 +6,8 @@
   number={4},
   pages={524--531},
   year={2003},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  DOI={10.1093/bioinformatics/btg015}
 }
 
 @article{Augustin:2023,
@@ -33,6 +34,7 @@
 	Title = {Preclinical pharmacokinetic / pharmacodynamic modeling and simulation in the pharmaceutical industry: an IQ consortium survey examining the current landscape},
 	Volume = {17},
 	Year = {2015},
+  DOI={10.1208/s12248-014-9716-2}
 }
 
 @article{MORGAN:2018,
@@ -99,7 +101,8 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   number={6},
   pages={1--9},
   year={2013},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  DOI={10.1038/psp.2013.24}
 }
 
 @article{hosseini2018gpkpdsim,
@@ -109,7 +112,8 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   volume={45},
   pages={259--275},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.1007/s10928-017-9562-9}
 }
 
 @article{sorzano2021scipion,
@@ -120,7 +124,8 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   number={7},
   pages={1169--1178},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.1007/s11095-021-03065-1}
 }
 
 @article{rackauckas2020accelerated,
@@ -129,7 +134,8 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   journal={BioRxiv},
   pages={2020--11},
   year={2020},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  DOI={10.1101/2020.11.28.402297}
 }
 
 @article{cole2014maximum,
@@ -140,7 +146,8 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   number={2},
   pages={252--260},
   year={2014},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  DOI={10.1093/aje/kwt245}
 }
 
 @article{gelman2014understanding,
@@ -151,7 +158,8 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   number={6},
   pages={997--1016},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.1007/s11222-013-9416-2}
 }
 
 @article{augustin2022treatment,
@@ -160,7 +168,8 @@ abstract = {In this review we present ways in which translational PK/PD modeling
   journal={bioRxiv},
   pages={2022--03},
   year={2022},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  DOI={10.1101/2022.03.19.483454}
 }
 
 @article{clerx2016myokit,
