+++ paper.md
@@ -20,9 +20,9 @@ bibliography: paper.bib
 
 # Summary
 
-[Chi](https://chi.readthedocs.io) is an easy-to-use, open source Python package for the modelling of pharmacokinetics & pharmacodynamics (PKPD). We provide two flexible interfaces to implement PKPD models: 1. an SBML interface, which implements PKPD models based on SBML file specifications [@hucka:2003]; and 2. a general purpose interface that allows users to implement their own, custom PKPD models using Python code. PKPD models instantiated from SBML files automatically implement the administration of custom dosing regimens and the evaluation of parameter sensitivities [@clerx2016myokit]. We also provide a simple framework to extend PKPD models to nonlinear mixed effects (NLME) models, making the simulation of inter-individual variability of treatment responses possible.
+[Chi](https://chi.readthedocs.io) is an open source Python package designed for the modelling of treatment responses, with support for implementation, simulation and inference. Supported treatment response models include pharmacokinetic & pharmacodynamic (PKPD) models, physiology-based pharmacokinetic (PBPK) models, quantitative systems pharmacology (QSP) models, and nonlinear mixed effects (NLME) models. The package provides two flexible interfaces to implement single-individual treatment response models: 1. an SBML interface, which implements models based on SBML file specifications [@hucka:2003]; and 2. a general purpose interface that allows users to implement their own, custom models using Python code. Models implemented using SBML files automatically implement routines to administer custom dosing regimens and to evaluate parameter sensitivities using the simulation engine [Myokit](http://myokit.org/) [@clerx2016myokit]. These single-individual treatment response models can then be extending to NLME models, making the simulation of inter-individual variability of treatment responses possible.
 
-In [Chi](https://chi.readthedocs.io), model parameters can be estimated from data using Bayesian inference. We provide a simple interface to estimate posterior distributions of PKPD model parameters and NLME model parameters. [Chi](https://chi.readthedocs.io) also implements filter inference, a novel inference approach which makes the estimation of NLME model parameters from snapshot time series data possible [@Augustin:2023].
+In [Chi](https://chi.readthedocs.io), model parameters can be estimated from data using Bayesian inference. We provide a simple interface to estimate posterior distributions of model parameters for individuals, e.g. PKPD model parameters, or for populations, i.e. NLME model parameters. [Chi](https://chi.readthedocs.io) also implements filter inference, a novel inference approach which makes the estimation of NLME model parameters from snapshot time series data possible [@Augustin:2023].
 
 For the sampling from posterior distributions, [Chi](https://chi.readthedocs.io) uses Markov chain Monte Carlo (MCMC) algorithms implemented in the Python package [PINTS](https://pints.readthedocs.io/en/stable/) [@Clerx:2019].
 
@@ -30,10 +30,10 @@ Documentation, tutorials and install instructions are available at https://chi.r
 
 # Statement of need
 
-PKPD modelling has become an integral part of pharmaceutical research [@SCHUCK:2015; @MORGAN:2018]. In the early phase of drug development, PKPD models help with target and lead identification, and contribute to a (semi-)mechanistic understanding of the relevant pharmacological processes. The modelling results thereafter provide guidance in the transition to the clinical development phase [@LAVE:2016]. During clinical trials, PKPD models help predict the safety, efficacy and treatment response variability of different dosing strategies. More recently, PKPD models are also considered to help with the individualisation of dosing regimens of otherwise difficult-to-administer drugs [@Augustin:20232].
[…]
