+++ paper.md
@@ -20,11 +20,11 @@ bibliography: paper.bib
 
 # Summary
 
-[Chi](https://chi.readthedocs.io) is an open source Python package designed for the modelling of treatment responses, with support for implementation, simulation and inference. Supported treatment response models include pharmacokinetic & pharmacodynamic (PKPD) models, physiology-based pharmacokinetic (PBPK) models, quantitative systems pharmacology (QSP) models, and nonlinear mixed effects (NLME) models. The package provides two flexible interfaces to implement single-individual treatment response models: 1. an SBML interface, which implements models based on SBML file specifications [@hucka:2003]; and 2. a general purpose interface that allows users to implement their own, custom models using Python code. Models implemented using SBML files automatically implement routines to administer custom dosing regimens and to evaluate parameter sensitivities using the simulation engine [Myokit](http://myokit.org/) [@clerx2016myokit]. These single-individual treatment response models can then be extending to NLME models, making the simulation of inter-individual variability of treatment responses possible.
+[Chi](https://chi.readthedocs.io) is an open source Python package designed for the modelling of treatment responses with support for implementation, simulation and inference. Supported treatment response models include pharmacokinetic & pharmacodynamic (PKPD) models, physiology-based pharmacokinetic (PBPK) models, quantitative systems pharmacology (QSP) models, and nonlinear mixed effects (NLME) models. The package provides two interfaces to implement single-individual treatment response models: 1. an SBML interface, which implements models based on SBML file specifications [@hucka:2003]; and 2. a general purpose interface that allows users to implement their own, custom models using Python code. Models implemented using SBML files automatically provide routines to administer dosing regimens and to evaluate parameter sensitivities using the simulation engine [Myokit](http://myokit.org/) [@clerx2016myokit]. These single-individual treatment response models can be extended to NLME models, making the simulation of inter-individual variability of treatment responses possible.
 
-In [Chi](https://chi.readthedocs.io), model parameters can be estimated from data using Bayesian inference. We provide a simple interface to estimate posterior distributions of model parameters for individuals, e.g. PKPD model parameters, or for populations, i.e. NLME model parameters. [Chi](https://chi.readthedocs.io) also implements filter inference, a novel inference approach which makes the estimation of NLME model parameters from snapshot time series data possible [@Augustin:2023].
+In [Chi](https://chi.readthedocs.io), model parameters can be estimated from data using Bayesian inference. We provide a simple interface to estimate posterior distributions of model parameters from single-patient data or from population data. The package also implements filter inference, a novel inference approach which makes the estimation of NLME model parameters from snapshot time series data possible [@Augustin:2023]. In the context of model-informed precision dosing, [Chi](https://chi.readthedocs.io) may also be used to optimise dosing regimens in order to target a desired treatment response.
 
-For the sampling from posterior distributions, [Chi](https://chi.readthedocs.io) uses Markov chain Monte Carlo (MCMC) algorithms implemented in the Python package [PINTS](https://pints.readthedocs.io/en/stable/) [@Clerx:2019].
+For the sampling from posterior distributions, [Chi](https://chi.readthedocs.io) uses Markov chain Monte Carlo (MCMC) algorithms implemented in the Python package [PINTS](https://pints.readthedocs.io/en/stable/) [@Clerx:2019]. For the optimisation of dosing regimens, different optimisers can be used, including optimisers implemented in SciPy or in [PINTS](https://pints.readthedocs.io/en/stable/) [@2020SciPy-NMeth,@Clerx:2019].
 
 Documentation, tutorials and install instructions are available at https://chi.readthedocs.io.
 
