+++ paper.md
@@ -34,6 +34,6 @@ Treatment response modelling has become an integral part of pharmaceutical resea
 
 The most widely used software packages and computer programs for treatment response modelling include NONMEM [@keizer2013modeling], [Monolix](https://lixoft.com/products/monolix/), and Matlab Simbiology [@hosseini2018gpkpdsim]. Other software packages include Scipion PKPD [@sorzano2021scipion], [PoPy](https://product.popypkpd.com/), Pumas [@rackauckas2020accelerated], and a number of [R libraries](https://cran.r-project.org/web/views/Pharmacokinetics.html). These packages provide an extensive toolkit for PKPD modelling. However, most of these solutions are difficult to use for research as their source code is not publicly distributed or not subject to open-source licenses, concealing the algorithmic details, limiting the transparency of the modelling results, and hindering the methdological development. Notable exceptions are Scipion PKPD and the R libraries, which make their source code publicly available on GitHub.
 
-[Chi](https://chi.readthedocs.io/en/latest/index.html) is an easy-to-use, open-source Python package for the modelling of treatment responses. It is targeted at modellers on all levels of programming expertise. Modellers with an interest in methodological research can use [Chi](https://chi.readthedocs.io/en/latest/index.html)'s modular and open source framework to scrutinise their modelling results and study the advantages and limitations of diffent modelling choices. We hope that [Chi](https://chi.readthedocs.io) is the beginning of a community effort where improved methodology for treatment response modelling is developed and shared publicly.
+[Chi](https://chi.readthedocs.io/en/latest/index.html) is an easy-to-use, open-source Python package for the modelling of treatment responses. It is targeted at modellers on all levels of programming expertise. Modellers with an interest in methodological research can use [Chi](https://chi.readthedocs.io/en/latest/index.html)'s modular and open source framework to scrutinise their modelling results and to study the advantages and limitations of different modelling choices. We hope that this open-source software package encourages a community effort where improved methodology for treatment response modelling is developed and shared publicly.
 
 # References
