### 2024-01-14 #5925 @DavAug
Thank you @ppxasjsm for all your help so far! I have no completed the outstanding tasks.

- The release of the software with the latest changes from the review has the version number: 1.0.0
- I have archived the release on Zenodo and the DOI is: [10.5281/zenodo.10510572](https://zenodo.org/doi/10.5281/zenodo.10510572)
