### 2024-02-03 f27635b
remove isbn entry from bib
