+++ paper.bib
@@ -45,7 +45,6 @@
 	Date-Modified = {2021-07-24 12:31:33 +0100},
 	Doi = {10.1038/nrd.2017.244},
 	Id = {Morgan2018},
-	Isbn = {1474-1784},
 	Journal = {Nature Reviews Drug Discovery},
 	Number = {3},
 	Pages = {167--181},
