+++ paper.md
@@ -32,8 +32,8 @@ Documentation, tutorials and install instructions are available at https://chi.r
 
 Treatment response modelling has become an integral part of pharmaceutical research [@SCHUCK:2015; @MORGAN:2018]. In the early phase of drug development, treatment response models help with target and lead identification, and contribute to a mechanistic understanding of the relevant pharmacological processes. In the transition to the clinical development phase, these models provide guidance and help to identify safe and efficacious dosing regimens [@LAVE:2016]. During clinical trials, treatment response models further facilitate the assessment of safety, efficacy and treatment response variability. More recently, treatment response models are also being used in the context of MIPD, where models help to predict individualised dosing regimens for otherwise difficult-to-administer drugs [@Augustin:20232].
 
-The most widely used software packages and computer programs for treatment response modelling include NONMEM [@keizer2013modeling], [Monolix](https://lixoft.com/products/monolix/), and Matlab Simbiology [@hosseini2018gpkpdsim]. Other software packages include Scipion PKPD [@sorzano2021scipion], [PoPy](https://product.popypkpd.com/), Pumas [@rackauckas2020accelerated], and a number of [R libraries](https://cran.r-project.org/web/views/Pharmacokinetics.html). These packages provide an extensive toolkit for PKPD modelling. However, most of these solutions are difficult to use for research as their source code is not publicly distributed or not subject to open-source licenses, concealing the algorithmic details, limiting the transparency of the modelling results, and hindering the methdological development. Notable exceptions are Scipion PKPD and the R libraries, which make their source code publicly available on GitHub.
+The most widely used software packages and computer programs for treatment response modelling include NONMEM [@keizer2013modeling], [Monolix](https://lixoft.com/products/monolix/), and Matlab Simbiology [@hosseini2018gpkpdsim]. Other software packages include Scipion PKPD [@sorzano2021scipion], [PoPy](https://product.popypkpd.com/), Pumas [@rackauckas2020accelerated], and a number of [R libraries](https://cran.r-project.org/web/views/Pharmacokinetics.html). These packages provide an extensive toolkit for PKPD modelling. However, most of these solutions are difficult to use for research as their source code is neither publicly distributed nor subject to open-source licenses, which conceals the algorithmic details, limits the transparency of the modelling results, and hinders the methdological development. Notable exceptions are Scipion PKPD and the R libraries, which make their source code publicly available on GitHub.
 
-[Chi](https://chi.readthedocs.io/en/latest/index.html) is an easy-to-use, open-source Python package for treatment response modelling. It is targeted at modellers on all levels of programming expertise. Modellers with a primary focus on the pharmacology can use [Chi](https://chi.readthedocs.io/en/latest/index.html) to quickly implement models and estimate their model parameters from data. Modellers with an interest in methodological research can use [Chi](https://chi.readthedocs.io/en/latest/index.html)'s modular, open source framework to study the advantages and limitations of different modelling choices, and research new approaches for treatment response modelling. We hope that the open-source nature of this package will increase the transparency of treatment response models and facilitate a community effort to further develop their methodology.
+[Chi](https://chi.readthedocs.io/en/latest/index.html) is an easy-to-use, open-source Python package for treatment response modelling. It is targeted at modellers on all levels of programming expertise. Modellers with a primary focus on the pharmacology can use [Chi](https://chi.readthedocs.io/en/latest/index.html) to quickly implement models and estimate their model parameters from data. Modellers with an interest in methodological research can use [Chi](https://chi.readthedocs.io/en/latest/index.html)'s modular, open source framework to study the advantages and limitations of different modelling choices, as well as research new approaches for treatment response modelling. We hope that the open-source nature of this package will increase the transparency of treatment response models and facilitate a community effort to further develop their methodology.
 
 # References
