### 2024-02-03 #5925 @DavAug
Hi @ppxasjsm @Kevin-Mattheus-Moerman

Thank you so much for recommending the acceptance of my article. I really appreciate it! :)

I just had a quick look into the error message, and while I am not sure why the error occured, I noticed that it was related to "wrong" ISBN in one of my BibTeX entries. Since the ISBN is not actually included in the reference that appear on the rendered article, I simply removed it.

@ppxasjsm would you, perhaps, be able to trigger the acceptance again? I hope that this will remove the error (I saw that an article was successfully published yesterday, so the error does seem to be specific to my article).

Thank you so much and best wishes,
David
