### 2024-02-06 #5925 @DavAug
> :wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
>
> Check final proof :point_right::page_facing_up: [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05925/joss.05925/10.21105.joss.05925.pdf)
>
> If the paper PDF and the deposit XML files look good in https://github.com/openjournals/joss-papers/pull/4979, then you can now move forward with accepting the submission by compiling again with the command `@editorialbot accept`

Hi @Kevin-Mattheus-Moerman , thank you so much for recommending the acceptance.

I just wanted to kindly confirm that accepting the submission is not in my mandate? I.e. we are not waiting for me to approve the final version of the article.

Best,
David
