### 2024-02-07 #5925 @Kevin-Mattheus-Moerman
@DavAug As AEiC I will now help process this submission for acceptance in JOSS. I have just checked your repository, the archive link, this review, and the paper. Most seems in order. I do however have the below points that require your attention:

- [ ] In your affiliation, please spell out `UK` as `United Kingdom`
- [ ] Please edit the ZENODO archive version tag to be `v1.0.0`
- [ ] Check `methdological` which should read `methodological`
