+++ CITATION.cff
@@ -1,5 +1,6 @@
 cff-version: 1.0.0
 message: "If you use this software, please cite it as below."
+type: "article"
 authors:
 - family-names: "Augustin"
   given-names: "David"
