+++ CITATION.cff
@@ -1,6 +1,6 @@
 cff-version: 1.0.0
 message: "If you use this software, please cite it as below."
-type: "article"
+preferred-citation: "article"
 authors:
 - family-names: "Augustin"
   given-names: "David"
