### 2024-02-08 ddf65bb
Update CITATION.cff
